add checks on AF sequence number and dlfc.
In particular, ignore frames with dlfc between last_dlfc - 10 and last_dlfc.

This will stop network generated duplicates causing problems.

+++ common/MDI/TagPacketDecoder.h
@@ -58,7 +58,7 @@ protected:
 	void InitTagItemDecoders(void);
 private:
 	CVector<CTagItemDecoder *> vecpTagItemDecoders;
-	int iSeqNumber;
+	uint16_t iSeqNumber;
 };
 
 #endif
